At last! Fixed certain characters breaking output
I've been thinking I'd do a big rewrite with JSON and Wunderbaum, but
that turned out to be time-consuming. So instead I just improved the
current template.

Fixed: Characters like asterisks and new lines no longer break the
output.
Fixed: Directories for which the current user does not have read
permissions will no longer break the output.

This means LinuxDir2HTML can finally index from / with --hidden and
still produce an accurate output. This is a pretty bad idea, especially
because /proc/kcore will be read as a file, but it CAN be done!

Finally, in some sense Firefox 33 and other contemporary browsers is
now a minimum requirement to browse the output due to changes in the
template.